Refactor voice controllers
